requirements.txt
flask~=1.1.2
wtforms~=2.3.3
pillow~=8.2.0
werkzeug~=1.0.1
itsdangerous~=2.0.1
alembic~=1.6.5
sqlalchemy~=1.4.15